moved css to separate file
